refactor(fp-frontend): flytt snarveg-miljøsjekk til paths
Flytt snarvegerErTilgjengelig til appens path-hjelpere og flytt testen fra snarveger-katalogen til paths.spec.ts.
