+++ apps/fp-frontend/src/app/paths.ts
@@ -6,6 +6,7 @@ import type { SkjermlenkeType, SkjermlenkeTypeFpTilbake } from '@navikt/fp-types
 
 export const UTBETALINGSDATA_PATH = 'utbetalingsdata';
 
+const PROD_HOSTNAME_DEL = 'intern.nav.no';
 const DEFAULT_FAKTA = 'default';
 const DEFAULT_PROSESS_STEG = 'default';
 
@@ -14,6 +15,9 @@ export const fagsakRoutePath = '/fagsak/:saksnummer//*';
 export const aktoerRoutePath = '/aktoer/:aktoerId';
 export const behandlingerRoutePath = '/behandling//*';
 export const behandlingRoutePath = '/:behandlingUuid/';
+
+export const snarvegerErTilgjengelig = (hostname = globalThis.location.hostname): boolean =>
+  !hostname.includes(PROD_HOSTNAME_DEL);
 
 export const pathToFagsak = (saksnummer: string): string => `/fagsak/${saksnummer}/`;
 export const pathToBehandlinger = (saksnummer: string): string => `/fagsak/${saksnummer}/behandling/`;
